Split apart examples
